Use Base paymaster on server side
Also extract creation of SmartAccountClient to be reused on server
